Implement base configuration for independent repository
Within the tox.ini under the [flake8] section the "max-line-length"
is being set because of the venv supported added to the #! at the top
of the python files which causes the tox linters check to fail. The
"max-line-length" has been set to 120.

Updates have been made to the swift sync python files so that they
pass lint checks.

Change-Id: Icab26bb57e2f5bd5d905080e583502aeeb423cf4
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>